third_party/WebKit/LayoutTests/media/video-pause-empty-events.html

<!DOCTYPE html>
<title>Test that pause() from EMPTY network state triggers load().</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = "content/test.ogv";
    assert_equals(video.networkState, HTMLMediaElement.NETWORK_NO_SOURCE);

    video.onloadstart = t.step_func(function() {});
    video.onratechange = t.step_func(function() {});
    video.onwaiting = t.step_func(function() {});
    video.onratechange = t.step_func(function() {});
    video.ondurationchange = t.step_func(function() {});
    video.onloadedmetadata = t.step_func(function() {});
    video.onloadeddata = t.step_func(function() {});
    video.oncanplay = t.step_func(function() {});
    video.oncanplaythrough = t.step_func_done();

    // TODO(srirama.m): Rewrite or remove this test as settings "src" itself triggers media load.
    video.pause();
});
</script>
